requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-jose[cryptography]
email-validator
pdfminer.six
